+++ tests/conftest.py
@@ -78,7 +78,7 @@ from great_expectations.self_check.util import (
     expectationSuiteValidationResultSchema,
     get_dataset,
 )
-from great_expectations.util import is_library_loadable
+from great_expectations.util import build_in_memory_runtime_context, is_library_loadable
 from tests.rule_based_profiler.parameter_builder.conftest import (
     RANDOM_SEED,
     RANDOM_STATE,
@@ -7022,56 +7022,6 @@ data_connectors:
             "name": "generic_csv_generator",
         }
     ]
-    return context
-
-
-def build_in_memory_runtime_context():
-    data_context_config: DataContextConfig = DataContextConfig(
-        datasources={
-            "pandas_datasource": {
-                "execution_engine": {
-                    "class_name": "PandasExecutionEngine",
-                    "module_name": "great_expectations.execution_engine",
-                },
-                "class_name": "Datasource",
-                "module_name": "great_expectations.datasource",
-                "data_connectors": {
-                    "runtime_data_connector": {
-                        "class_name": "RuntimeDataConnector",
-                        "batch_identifiers": [
-                            "id_key_0",
-                            "id_key_1",
-                        ],
-                    }
-                },
-            },
-            "spark_datasource": {
-                "execution_engine": {
-                    "class_name": "SparkDFExecutionEngine",
-                    "module_name": "great_expectations.execution_engine",
-                },
-                "class_name": "Datasource",
-                "module_name": "great_expectations.datasource",
-                "data_connectors": {
-                    "runtime_data_connector": {
-                        "class_name": "RuntimeDataConnector",
-                        "batch_identifiers": [
-                            "id_key_0",
-                            "id_key_1",
-                        ],
-                    }
-                },
-            },
-        },
-        expectations_store_name="expectations_store",
-        validations_store_name="validations_store",
-        evaluation_parameter_store_name="evaluation_parameter_store",
-        checkpoint_store_name="checkpoint_store",
-        store_backend_defaults=InMemoryStoreBackendDefaults(),
-    )
-
-    context: BaseDataContext = BaseDataContext(project_config=data_context_config)
-
     return context
